fix: removed redundant type file
