+++ src/components/Menu/Menu.types.ts
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import { ListProps } from '../List';
 import {
   MenuItemButtonProps,
@@ -8,7 +8,6 @@ import {
   DropdownMenuItemProps,
 } from './MenuItem/MenuItem.types';
 import { ButtonProps } from '../Button';
-import { MenuRenderFunction } from './Menu.shared.types';
 
 export enum MenuVariant {
   disruptive = 'disruptive',
@@ -95,4 +94,15 @@ export interface DropdownMenuProps extends MenuProps {
    */
   referenceOnKeydown?: (event: React.KeyboardEvent) => void;
   menuRenderer?: MenuRenderFunction;
+}
+export interface MenuRendererProps {
+  items: MenuItemTypes[];
+  size?: MenuSize;
+  variant?: MenuVariant;
+  children?: ReactNode;
+  [key: string]: any;
+}
+
+export interface MenuRenderFunction {
+  (props: MenuRendererProps): JSX.Element;
 }
